More refactoring progress
