+++ lib/drives/index.js
@@ -4,8 +4,9 @@ const pump = require('pump')
 const sub = require('subleveldown')
 const bjson = require('buffer-json-encoding')
 const collectStream = require('stream-collector')
-const Nanoresource = require('nanoresource-promise/emitter')
+const { NanoresourcePromise: Nanoresource } = require('nanoresource-promise/emitter')
 const { Transform } = require('streamx')
+
 
 const {
   fromHyperdriveOptions,
@@ -42,12 +43,10 @@ class DriveManager extends Nanoresource {
     this.watchLimit = opts.watchLimit
     this.memoryOnly = !!opts.memoryOnly
 
-    const dbs = DriveManager.SUB_DBS.reduce((acc, dbInfo) => {
-      acc[dbInfo.prefix] = sub(this.db, dbInfo.prefix, { valueEncoding: dbInfo.valueEncoding })
-    })
+    const dbs = DriveManager.generateSubDbs(db)
 
     this._driveIndex = dbs.drives
-    this._mirrorIndex = dbs.mirroring
+    this._mirrorIndex = dbs.mirrors
 
     this._drives = new Map()
     this._checkouts = new Map()
@@ -1019,18 +1018,12 @@ class DriveManager extends Nanoresource {
     })
   }
 }
-DriveManager.SUB_DBS = {
-  drives: {
-    prefix: 'drives',
-    valueEncoding: 'bjson'
-  },
-  mirroring: {
-    prefix: 'mirrors',
-    valueEncoding: 'utf8'
-  },
-  seeding: {
-    prefix: 'seeding',
-    valueEncoding: 'json'
+
+DriveManager.generateSubDbs = function (db) {
+  return {
+    drives: sub(db, 'drives', { valueEncoding: 'bjson' }),
+    mirrors: sub(db, 'mirrors', { valueEncoding: 'utf8' }),
+    seeding: sub(db, 'seeding', { valueEncoding: 'json '})
   }
 }
