+++ lib/network.js
@@ -1,7 +1,7 @@
 const sub = require('subleveldown')
 const bjson = require('buffer-json-encoding')
 const datEncoding = require('dat-encoding')
-const Nanoresource = require('nanoresource-promise/emitter')
+const { NanoresourcePromise: Nanoresource } = require('nanoresource-promise/emitter')
 
 const {
   fromNetworkConfiguration,
@@ -10,7 +10,7 @@ const {
 const { rpc } = require('hyperdrive-daemon-client')
 
 const { dbGet, dbCollect } = require('./common')
-const log = require('../log').child({ component: 'network-manager' })
+const log = require('./log').child({ component: 'network-manager' })
 
 class NetworkManager extends Nanoresource {
   constructor (networking, db, opts = {}) {
@@ -19,7 +19,8 @@ class NetworkManager extends Nanoresource {
     this.db = db
     this.noAnnounce = !!opts.noAnnounce
 
-    this._seedIndex = sub(this.db, 'seeding', { valueEncoding: 'json' })
+    const dbs = NetworkManager.generateSubDbs(db)
+    this._seedIndex = dbs.seeding
   }
 
   ready () {
@@ -34,7 +35,7 @@ class NetworkManager extends Nanoresource {
 
   async _rejoin () {
     if (this.noAnnounce) return
-    const seedList = await collect(this._seedIndex)
+    const seedList = await dbCollect(this._seedIndex)
     for (const { key: discoveryKey, value: networkOpts } of seedList) {
       const opts = networkOpts && networkOpts.opts
       if (!opts || !opts.announce) continue
@@ -112,10 +113,9 @@ class NetworkManager extends Nanoresource {
     return rsp
   }
 }
-NetworkManager.SUB_DBS = {
-  seeding: {
-    prefix: 'seeding',
-    valueEncoding: 'json'
+NetworkManager.generateSubDbs = function (db) {
+  return {
+    seeding: sub(db, 'seeding', { valueEncoding: 'json '})
   }
 }
