+++ lib/peers.js
@@ -5,7 +5,7 @@ const { rpc } = require('hyperdrive-daemon-client')
 const messages = rpc.peers.messages
 const WatchPeersTypes = messages.WatchPeersResponse.Type
 
-const log = require('../log').child({ component: 'peers' })
+const log = require('./log').child({ component: 'peers' })
 
 const ALIAS = Symbol('hyperdrive-peer-alias')
