+++ lib/peersockets.js
@@ -2,7 +2,7 @@ const { EventEmitter } = require('events')
 
 const { rpc } = require('hyperdrive-daemon-client')
 const messages = rpc.peersockets.messages
-const log = require('../log').child({ component: 'peersockets' })
+const log = require('./log').child({ component: 'peersockets' })
 
 const PeerMessageTypes = messages.PeerMessage.Type
